+++ src/client/rdfjs-buffer/mod.ts
@@ -3,3 +3,4 @@ export * from "./apply-rdfjs-patch-to-store.ts";
 export * from "./commit-buffered-patch.ts";
 export * from "./import-commit-target.ts";
 export * from "./import-export-via-rdfjs-store.ts";
+export * from "./buffered-rdfjs-quad-store.ts";